# contracts/salt-control-api/openapi.yaml
components:
  schemas:
...
    ConfigurePayload:
      properties:
        configRevision:
          anyOf:
          - type: string
          - type: 'null'
          title: Configrevision
        desired:
          additionalProperties: true
          title: Desired
          type: object
        kind:
          const: configure
          default: configure
          title: Kind
          type: string
      title: ConfigurePayload
      type: object
...
    DeviceInfo:
      properties:
        arch:
          title: Arch
          type: string
        hostname:
          title: Hostname
          type: string
        machineGuidHash:
          title: Machineguidhash
          type: string
        windowsBuild:
          title: Windowsbuild
          type: integer
      required:
      - hostname
      - machineGuidHash
      - windowsBuild
      - arch
      title: DeviceInfo
      type: object
...
    EnrollmentCreateRequest:
      properties:
        device:
          $ref: '#/components/schemas/DeviceInfo'
        enrollmentToken:
          title: Enrollmenttoken
          type: string
        requestId:
          title: Requestid
          type: string
      required:
      - enrollmentToken
      - requestId
      - device
      title: EnrollmentCreateRequest
      type: object
    EnrollmentCreateResponse:
      properties:
        deviceCredential:
          title: Devicecredential
          type: string
        endpointId:
          title: Endpointid
          type: string
        enrollmentId:
          title: Enrollmentid
          type: string
        expiresAt:
          title: Expiresat
          type: string
        masterFingerprints:
          items:
            type: string
          title: Masterfingerprints
          type: array
        masters:
          items:
            type: string
          title: Masters
          type: array
      required:
      - enrollmentId
      - endpointId
      - masters
      - masterFingerprints
      - deviceCredential
      - expiresAt
      title: EnrollmentCreateResponse
      type: object
    EnrollmentStatusResponse:
      properties:
        endpointId:
          title: Endpointid
          type: string
        enrollmentId:
          title: Enrollmentid
          type: string
        errorCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Errorcode
        masterFingerprints:
          items:
            type: string
          title: Masterfingerprints
          type: array
        masters:
          items:
            type: string
          title: Masters
          type: array
        state:
          title: State
          type: string
      required:
      - enrollmentId
      - endpointId
      - state
      - masters
      - masterFingerprints
      title: EnrollmentStatusResponse
      type: object
    FingerprintReportRequest:
      properties:
        endpointId:
          title: Endpointid
          type: string
        minionFingerprint:
          title: Minionfingerprint
          type: string
        requestId:
          title: Requestid
          type: string
      required:
      - endpointId
      - minionFingerprint
      - requestId
      title: FingerprintReportRequest
      type: object
    FingerprintReportResponse:
      properties:
        endpointId:
          title: Endpointid
          type: string
        enrollmentId:
          title: Enrollmentid
          type: string
        errorCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Errorcode
        state:
          title: State
          type: string
      required:
      - enrollmentId
      - endpointId
      - state
      title: FingerprintReportResponse
      type: object
    GatewayLifecyclePayload:
      properties:
        action:
          default: restart
          enum:
          - start
          - stop
          - restart
          title: Action
          type: string
        kind:
          const: gateway
          default: gateway
          title: Kind
          type: string
      title: GatewayLifecyclePayload
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HandoverPayload:
      properties:
        configRevision:
          anyOf:
          - type: string
          - type: 'null'
          title: Configrevision
        endpointId:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpointid
        kind:
          const: handover
          default: handover
          title: Kind
          type: string
        releaseId:
          anyOf:
          - type: string
          - type: 'null'
          title: Releaseid
      title: HandoverPayload
      type: object
    InstallPayload:
      properties:
        artifactUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Artifacturl
        component:
          default: hermes
          title: Component
          type: string
        kind:
          const: install
          default: install
          title: Kind
          type: string
        sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Sha256
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
      title: InstallPayload
      type: object
    JobCreateRequest:
      properties:
        configRevision:
          anyOf:
          - type: string
          - type: 'null'
          title: Configrevision
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        endpointId:
          title: Endpointid
          type: string
        idempotencyKey:
          title: Idempotencykey
          type: string
        minionId:
          title: Minionid
          type: string
        operation:
          enum:
          - install
          - configure
          - start
          - stop
          - restart
          - health
          - diagnose
          - rollback
          - handover
          - remigrate
          title: Operation
          type: string
        payload:
          anyOf:
          - discriminator:
              mapping:
                configure: '#/components/schemas/ConfigurePayload'
                gateway: '#/components/schemas/GatewayLifecyclePayload'
                handover: '#/components/schemas/HandoverPayload'
                install: '#/components/schemas/InstallPayload'
                probe: '#/components/schemas/ProbePayload'
                remigrate: '#/components/schemas/RemigratePayload'
                rollback: '#/components/schemas/RollbackPayload'
                upgrade: '#/components/schemas/UpgradePayload'
              propertyName: kind
            oneOf:
            - $ref: '#/components/schemas/InstallPayload'
            - $ref: '#/components/schemas/UpgradePayload'
            - $ref: '#/components/schemas/ConfigurePayload'
            - $ref: '#/components/schemas/GatewayLifecyclePayload'
            - $ref: '#/components/schemas/ProbePayload'
            - $ref: '#/components/schemas/HandoverPayload'
            - $ref: '#/components/schemas/RollbackPayload'
            - $ref: '#/components/schemas/RemigratePayload'
          - type: 'null'
          title: Payload
        releaseId:
          anyOf:
          - type: string
          - type: 'null'
          title: Releaseid
        requestedBy:
          title: Requestedby
          type: string
      required:
      - endpointId
      - minionId
      - operation
      - idempotencyKey
      - requestedBy
      title: JobCreateRequest
      type: object
    JobResponse:
      properties:
        acceptedAt:
          title: Acceptedat
          type: string
        conflictJobId:
          anyOf:
          - type: string
          - type: 'null'
          title: Conflictjobid
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        duplicate:
          default: false
          title: Duplicate
          type: boolean
        endpointId:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpointid
        errorCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Errorcode
        jobId:
          title: Jobid
          type: string
        minionId:
          anyOf:
          - type: string
          - type: 'null'
          title: Minionid
        operation:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation
        saltJid:
          anyOf:
          - type: string
          - type: 'null'
          title: Saltjid
        status:
          enum:
          - queued
          - dispatching
          - running
          - succeeded
          - failed
          - cancelled
          - expired
          title: Status
          type: string
      required:
      - jobId
      - status
      - acceptedAt
      title: JobResponse
      type: object
...
    ProbePayload:
      properties:
        kind:
          const: probe
          default: probe
          title: Kind
          type: string
        probe:
          default: health
          enum:
          - health
          - diagnose
          title: Probe
          type: string
      title: ProbePayload
      type: object
    RemigratePayload:
      properties:
        endpointId:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpointid
        idempotencyKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotencykey
        kind:
          const: remigrate
          default: remigrate
          title: Kind
          type: string
      title: RemigratePayload
      type: object
...
    RollbackPayload:
      properties:
        kind:
          const: rollback
          default: rollback
          title: Kind
          type: string
        previousOwner:
          anyOf:
          - type: string
          - type: 'null'
          title: Previousowner
      title: RollbackPayload
      type: object
...
    UpgradePayload:
      properties:
        artifactUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Artifacturl
        kind:
          const: upgrade
          default: upgrade
          title: Kind
          type: string
        sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Sha256
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
      title: UpgradePayload
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: SMC Salt Control
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /salt/v1/enrollments:
    post:
      operationId: create_enrollment_salt_v1_enrollments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentCreateResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Enrollment
      tags:
      - enrollments
  /salt/v1/enrollments/{enrollment_id}:
    get:
      operationId: get_enrollment_salt_v1_enrollments__enrollment_id__get
      parameters:
      - in: path
        name: enrollment_id
        required: true
        schema:
          title: Enrollment Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Enrollment
      tags:
      - enrollments
  /salt/v1/enrollments/{enrollment_id}/fingerprint:
    post:
      operationId: report_fingerprint_salt_v1_enrollments__enrollment_id__fingerprint_post
      parameters:
      - in: path
        name: enrollment_id
        required: true
        schema:
          title: Enrollment Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FingerprintReportRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FingerprintReportResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Report Fingerprint
      tags:
      - enrollments
...
  /salt/v1/migrations/handover:
    post:
      operationId: handover_salt_v1_migrations_handover_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Handover
      tags:
      - migrations
  /salt/v1/migrations/remigrate:
    post:
      operationId: remigrate_salt_v1_migrations_remigrate_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Remigrate
      tags:
      - migrations
  /salt/v1/migrations/rollback:
    post:
      operationId: rollback_salt_v1_migrations_rollback_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rollback
      tags:
      - migrations
...